[AF-485] feat (FRONT posts): refactor directory
