+++ front/src/app/app-routing.module.ts
@@ -1,9 +1,9 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { HomeComponent } from './features/auth/components/home/home.component';
-import { UnauthGuard } from './guards/unauth.guard';
-import { AuthGuard } from './guards/auth.guard';
-import { UserDetailsComponent } from './features/feed/components/user-details/user-details.component';
+import { UnauthGuard } from './core/guards/unauth.guard';
+import { AuthGuard } from './core/guards/auth.guard';
+import { TopicsComponent } from './features/topics/topics.component';
+import { UserDetailsComponent } from './features/user-details/user-details.component';
 
 const routes: Routes = [
   {
@@ -13,10 +13,20 @@ const routes: Routes = [
       import('./features/auth/auth.module').then((m) => m.AuthModule),
   },
   {
-    path: 'feed',
+    path: 'post',
     canActivate: [AuthGuard],
     loadChildren: () =>
-      import('./features/feed/feed.module').then((m) => m.FeedModule),
+      import('./features/post/post.module').then((m) => m.PostModule),
+  },
+  {
+    path: 'topics',
+    canActivate: [AuthGuard],
+    component: TopicsComponent,
+  },
+  {
+    path: 'user-details',
+    canActivate: [AuthGuard],
+    component: UserDetailsComponent,
   },
 ];
