+++ front/src/app/core/interceptors/jwt.interceptor.ts
@@ -1,5 +1,9 @@
-import { HttpHandler, HttpInterceptor, HttpRequest } from "@angular/common/http";
-import { Injectable } from "@angular/core";
+import {
+  HttpHandler,
+  HttpInterceptor,
+  HttpRequest,
+} from '@angular/common/http';
+import { Injectable } from '@angular/core';
 import { SessionService } from '../services/session.service';
 
 @Injectable({ providedIn: 'root' })
@@ -10,7 +14,9 @@ export class JwtInterceptor implements HttpInterceptor {
     if (this.sessionService.isLogged) {
       request = request.clone({
         setHeaders: {
-          Authorization: `Bearer ${this.sessionService.sessionInformation!.token}`,
+          Authorization: `Bearer ${
+            this.sessionService.sessionInformation!.token
+          }`,
         },
       });
     }
