+++ front/src/app/core/interfaces/loginRequest.interface.ts
