+++ front/src/app/core/interfaces/post.interface.ts
