+++ front/src/app/core/interfaces/registerRequest.interface.ts
