+++ front/src/app/core/interfaces/sessionInformation.interface.ts
