+++ front/src/app/core/interfaces/topic.interface.ts
