+++ front/src/app/core/interfaces/user.interface.ts
