+++ front/src/app/core/services/topic.service.ts
