+++ front/src/app/core/services/user.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 
-import { UserService } from './user.service';
+import { UserService } from '../features/feed/services/user.service';
 
 describe('UserService', () => {
   let service: UserService;
