+++ front/src/app/features/auth/register/register.component.spec.ts
