--- front/src/app/features/feed/feed.module.ts
@@ -1,14 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-
-import { FeedRoutingModule } from './feed-routing.module';
-
-
-@NgModule({
-  declarations: [],
-  imports: [
-    CommonModule,
-    FeedRoutingModule
-  ]
-})
-export class FeedModule { }
