+++ front/src/app/features/post/components/feed/feed.component.html
