+++ front/src/app/features/post/post-routing.module.ts
@@ -1,0 +1,13 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import { FeedComponent } from './components/feed/feed.component';
+
+const routes: Routes = [
+  { title: 'Feed', path: 'feed', component: FeedComponent },
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule],
+})
+export class PostRoutingModule {}
