+++ front/src/app/features/post/post.module.ts
@@ -1,7 +1,8 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { AuthRoutingModule } from './auth-routing.module';
+import { PostRoutingModule } from './post-routing.module';
+import { FeedComponent } from './components/feed/feed.component';
 import { MatButtonModule } from '@angular/material/button';
 import { MatCardModule } from '@angular/material/card';
 import { MatIconModule } from '@angular/material/icon';
@@ -9,10 +10,12 @@ import { MatSnackBarModule } from '@angular/material/snack-bar';
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
-import { RegisterComponent } from './register/register.component';
-import { LoginComponent } from './login/login.component';
+import { MatSidenavModule } from '@angular/material/sidenav';
+import { MatListModule } from '@angular/material/list';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { LayoutModule } from '@angular/cdk/layout';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { HomeComponent } from './home/home.component';
+import { HttpClientModule } from '@angular/common/http';
 
 const materialModule = [
   MatButtonModule,
@@ -22,16 +25,18 @@ const materialModule = [
   MatToolbarModule,
   MatFormFieldModule,
   MatInputModule,
+  MatSidenavModule,
+  MatListModule,
 ];
 
 @NgModule({
-  declarations: [RegisterComponent, LoginComponent, HomeComponent],
+  declarations: [FeedComponent],
   imports: [
-    CommonModule,
-    AuthRoutingModule,
+    PostRoutingModule,
+    LayoutModule,
     FormsModule,
     ReactiveFormsModule,
     ...materialModule,
   ],
 })
-export class AuthModule {}
+export class PostModule {}
