+++ front/src/app/features/user-details/user-details.component.html
